Support basic HTTP get file request (#254)
* * Add newline header to startup (it sometimes appears on a continuation line)
* Print SD card directory on startup
* Add all serial output to the SD card as date.log

* Sample requests:
- http://192.168.1.156/echo?value="foo"
- http://192.168.1.156/20210825.log

* spelling

* more spelling

* spelling expects 'yy'

* * Use `int` for freeMemory result.
* Add some commented-out debugging logging.

* Merge from main with passing tests.

* Reduce memory usage.

* Tests pass.

* Use larger buffer to send files.

* remove commented-out code

* Address some Codacy warnings.

+++ src/Devices/PushingBox.h
@@ -26,7 +26,7 @@ private:
   EthernetClient client;
   const char *DevID = nullptr;  // DeviceID assigned by PushingBox and passed-in from TankController.ino
   // wait a bit for first reading (https://github.com/Open-Acidification/TankController/issues/179)
-  unsigned long nextSendTime = 70000;
+  uint32_t nextSendTime = 70000;
   const char *server = "api.pushingbox.com";
 
   // instance method
